JAMES-3023 Move EncryptionCodec to backends-cassandra for reuse
