+++ backends-common/cassandra/src/main/java/org/apache/james/backends/cassandra/encryption/EncryptionConfiguration.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.blob.cassandra.encryption;
+package org.apache.james.backends.cassandra.encryption;
 
 import java.util.Optional;
