+++ backends-common/cassandra/src/test/java/org/apache/james/backends/cassandra/encryption/EncryptionCodecTest.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.blob.cassandra.encryption;
+package org.apache.james.backends.cassandra.encryption;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
